+++ packages/api/src/__generated__/schema.ts
@@ -621,6 +621,8 @@ export type ProfileAddress = {
 
 export type Query = {
   __typename?: 'Query';
+  /** Returns the account name of the current user or the B2B contract name if applicable. */
+  accountName?: Maybe<Scalars['String']>;
   /** Returns information about all collections. */
   allCollections: StoreCollectionConnection;
   /** Returns information about all products. */