+++ packages/core/src/components/search/SearchInput/SearchInput.tsx
@@ -109,7 +109,6 @@ const SearchInput = forwardRef<SearchInputRef, SearchInputProps>(
       addToSearchHistory({ term, path })
       sendAnalytics(term)
       setSearchDropdownVisible(false)
-      setSearchQuery(term)
     }
 
     useOnClickOutside(searchRef, () =>