+++ packages/core/test/server/index.test.ts
@@ -77,6 +77,7 @@ const QUERIES = [
   'userOrder',
   'listUserOrders',
   'pickupPoints',
+  'accountName',
 ]
 
 const MUTATIONS = [